refactor with appTitle
