--- app/styleguide/index.scss
@@ -1,5 +1,0 @@
-/**
- * Style Guide
- * those styles are specific to the style guide project
- */
-
