+++ reapp-dev-tools/src/index.js
@@ -1,68 +1,15 @@
 
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { renderMultiComponents } from './render-multi-components';
+import { renderSingleComponent } from './render-single-component';
+import { renderStyleguideInfo } from './render-styleguide-info';
 
-import { Styleguide } from './Styleguide';
-import { StyleguideTitle } from './StyleguideTitle';
-import { StyleguideToc } from './StyleguideToc';
-import { StyleguidePage } from './StyleguidePage';
-import { StyleguideError } from './StyleguideError';
+import { SGPage } from './SGPage';
+import { SGSection } from './SGSection';
 
 module.exports = {
-    SGPage: require('./SGPage').SGPage,
-    SGSection: require('./SGSection').SGSection,
+    SGPage,
+    SGSection,
     renderMultiComponents,
     renderSingleComponent,
     renderStyleguideInfo,
 };
-
-function renderMultiComponents(targetEl, root, components, sources) {
-    try {
-        var GuideComponent;
-        var pages = components.map(component => {
-            GuideComponent = component.def;
-            return (
-                <StyleguidePage
-                    name={component.name}
-                    key={component.name}
-                    component={GuideComponent} />
-            );
-        });
-
-        ReactDOM.render((
-            <Styleguide root={root} sources={sources}>
-                <StyleguideTitle title="es6-react-client" />
-                <StyleguideToc components={components} />
-                {pages}
-            </Styleguide>
-        ), targetEl);
-
-        return true;
-
-    } catch (e) {
-        return e;
-    }
-}
-
-function renderSingleComponent(targetEl, root, component, sources) {
-    try {
-        var GuideComponent = component.def;
-        ReactDOM.render((
-            <Styleguide root={root} sources={sources}>
-                <StyleguideTitle />
-                <StyleguidePage
-                    name={component.name}
-                    component={GuideComponent} />
-            </Styleguide>
-        ), targetEl);
-    } catch (e) {
-        return e;
-    }
-    return true;
-}
-
-function renderStyleguideInfo(targetEl, root, error) {
-    ReactDOM.render((
-        <StyleguideError root={root} error={error} />
-    ), targetEl);
-}
